requirements.txt
Django==4.1.3
djangorestframework==3.14.0
ipython==8.6.0
Pillow==9.3.0
psycopg2==2.9.5
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
